Separate into "fixflag".
Addresses in issue #2.
