+++ single/fixflag.hpp
@@ -1,0 +1,10 @@
+/*******************************************************/
+/** 「空白マス非存在性」に基づきフラグセットする関数  **/
+/*******************************************************/
+
+#ifndef __FIXFLAG_HPP__
+#define __FIXFLAG_HPP__
+
+void generateFixFlag();
+
+#endif
